[client] Refactor User page
