+++ client/src/pages/User/index.tsx
@@ -1,18 +1,6 @@
 // Antd dependencies
-import { Card, Row, Col, Avatar, Typography, Tabs, Button, Pagination, Divider, Empty, Tooltip } from 'antd'
-import {
-	EditOutlined,
-	DownOutlined,
-	UpOutlined,
-	CopyOutlined,
-	LoadingOutlined,
-	SettingOutlined,
-	IdcardOutlined,
-	UpSquareOutlined,
-	LinkOutlined,
-	SolutionOutlined,
-	MessageOutlined,
-} from '@ant-design/icons'
+import { Card, Row, Col, Avatar, Typography, Divider, Tooltip } from 'antd'
+import { SettingOutlined, IdcardOutlined, UpSquareOutlined, LinkOutlined, SolutionOutlined, MessageOutlined } from '@ant-design/icons'
 import { GridContent } from '@ant-design/pro-layout'
 
 // Other dependencies
@@ -21,25 +9,18 @@ import { useSelector } from 'react-redux'
 import { history } from 'umi'
 
 // Local files
-import { fetchAllEntriesByAuthor, fetchUserByUsername, fetchUserVotes, fetchAllFeeds } from '@/services/api'
+import { fetchUserByUsername } from '@/services/api'
 import { API_URL } from '../../../config/constants'
 import { PageHelmet } from '@/components/PageHelmet'
 import PageLoading from '@/components/PageLoading'
 import NotFoundPage from '../404'
+import { UserTabs } from './components/UserTabs'
 
 const User: React.FC = ({ match }): JSX.Element => {
 	const userState = useSelector((state: any) => state.user?.attributes.user)
 
 	const [user, setUser] = useState(userState)
 	const [isUserFound, setIsUserFound] = useState<boolean | null>(null)
-
-	const [currentTab, setCurrentTab] = useState('feeds')
-	const [totalItems, setTotalItems] = useState<number>(0)
-	const [tabView, setTabView] = useState<JSX.Element>(
-		<div style={{ textAlign: 'center' }}>
-			<LoadingOutlined style={{ fontSize: 20 }} />
-		</div>
-	)
 
 	const readableRoles = {
 		0: 'User',
@@ -62,177 +43,6 @@ const User: React.FC = ({ match }): JSX.Element => {
 		}
 		setIsUserFound(true)
 	}, [])
-
-	const handleFeedsTabView = (skip: number): void => {
-		fetchAllFeeds(skip, match.params.username, undefined, undefined).then(async res => {
-			setTotalItems(res.data.attributes.count)
-
-			if (res.data.attributes.titles.length === 0) {
-				setTabView(
-					<div style={{
-						minHeight: 460,
-						display: 'flex',
-						alignItems: 'center',
-						justifyContent: 'center'
-					}}>
-						<Empty description="No Feed Found" image={Empty.PRESENTED_IMAGE_SIMPLE} />
-					</div>
-				)
-				return
-			}
-
-			const feedList = await res.data.attributes.titles.map((title: any) =>
-				<Row key={title.id} style={{ padding: 5 }}>
-					<Col>
-						{title.name}
-					</Col>
-					<Button onClick={(): void => history.push(`/${title.slug}`)} size="small" type="primary">
-						Open
-					</Button>
-				</Row>
-			)
-			setTabView(
-				<div style={{ minHeight: 460 }}>
-					{feedList}
-				</div>
-			)
-		})
-	}
-
-	const handleEntriesTabView = (skip: number): void => {
-		fetchAllEntriesByAuthor(match.params.username, skip).then(async res => {
-			setTotalItems(res.data.attributes.count)
-
-			if (res.data.attributes.entries.length === 0) {
-				setTabView(
-					<div style={{
-						minHeight: 460,
-						display: 'flex',
-						alignItems: 'center',
-						justifyContent: 'center'
-					}}>
-						<Empty description="No Entry Found" image={Empty.PRESENTED_IMAGE_SIMPLE} />
-					</div>
-				)
-				return
-			}
-
-			const entryList = await res.data.attributes.entries.map((entry) =>
-				<Row key={entry.id} style={{ padding: 5 }}>
-					<Col>
-						<Typography.Paragraph ellipsis>
-							{entry.text}
-						</Typography.Paragraph>
-					</Col>
-					<Button onClick={(): void => history.push(`/entry/${entry.id}`)} size="small" type="primary">
-						Open
-					</Button>
-				</Row>
-			)
-			setTabView(
-				<div style={{ minHeight: 460 }}>
-					{entryList}
-				</div>
-			)
-		})
-	}
-
-	const handleVotesTabView = (voteType: 'up' | 'down', skip: number): void => {
-		fetchUserVotes(match.params.username, voteType, skip).then(async res => {
-			setTotalItems(res.data.attributes.count)
-
-			if (res.data.attributes.entries.length === 0) {
-				setTabView(
-					<div style={{
-						minHeight: 460,
-						display: 'flex',
-						alignItems: 'center',
-						justifyContent: 'center'
-					}}>
-						<Empty
-							description={`This user has not ${voteType} vote any entry yet`}
-							image={Empty.PRESENTED_IMAGE_SIMPLE}
-						/>
-					</div>
-				)
-				return
-			}
-
-			const entryList = await res.data.attributes.entries.map((entry: any) =>
-				<Row key={entry.id} style={{ padding: 5 }}>
-					<Col>
-						<Typography.Paragraph ellipsis>
-							{entry.text}
-						</Typography.Paragraph>
-					</Col>
-					<Button onClick={(): void => history.push(`/entry/${entry.id}`)} size="small" type="primary">
-						Open
-					</Button>
-				</Row>
-			)
-			setTabView(
-				<div style={{ minHeight: 460 }}>
-					{entryList}
-				</div>
-			)
-		})
-	}
-
-	const handleLoadingView = (
-		<div style={{
-			minHeight: 460,
-			display: 'flex',
-			alignItems: 'center',
-			justifyContent: 'center'
-		}}>
-			<LoadingOutlined style={{ fontSize: 20 }} />
-		</div>
-	)
-
-	const handlePagination = (func: Function, voteType?: 'up' | 'down'): JSX.Element => (
-		<Pagination
-			size="small"
-			showLessItems={true}
-			showQuickJumper={false}
-			defaultCurrent={1}
-			pageSize={10}
-			hideOnSinglePage
-			total={totalItems}
-			onChange={
-				voteType ?
-					(page: number): void => {
-						setTabView(handleLoadingView)
-						func(voteType, 10 * ( page -1))
-					}
-				:
-					(page: number): void => {
-						setTabView(handleLoadingView)
-						func(10 * ( page -1))
-					}
-			}
-		/>
-	)
-
-	const handleTabChange = (key: string): void => setCurrentTab(key)
-
-	useEffect(() => {
-		setTabView(handleLoadingView)
-		setTotalItems(0)
-		switch (currentTab) {
-			case 'feeds':
-				handleFeedsTabView(0)
-				break
-			case 'entries':
-				handleEntriesTabView(0)
-				break
-			case 'up-votes':
-				handleVotesTabView('up', 0)
-				break
-			case 'down-votes':
-				handleVotesTabView('down', 0)
-				break
-		}
-	}, [currentTab])
 
 	if (isUserFound === null) return <PageLoading />
 	if (isUserFound === false) return <NotFoundPage />
@@ -328,74 +138,7 @@ const User: React.FC = ({ match }): JSX.Element => {
 						</Card>
 					</Col>
 					<Col lg={14} md={24}>
-						<Card bordered={false}>
-							<Tabs size="small" animated={false} onChange={handleTabChange} defaultActiveKey={currentTab}>
-								<Tabs.TabPane
-									tab={
-										<Typography.Text strong>
-											<CopyOutlined style={{ marginRight: 0 }} /> Created Feeds
-										</Typography.Text>
-									}
-									key="feeds"
-								>
-									{tabView}
-									{totalItems > 0 &&
-										<>
-											<Divider />
-											{handlePagination(handleFeedsTabView)}
-										</>
-									}
-								</Tabs.TabPane>
-								<Tabs.TabPane
-									tab={
-										<Typography.Text strong>
-											<EditOutlined style={{ marginRight: 0 }} /> Created Entries
-										</Typography.Text>
-									}
-									key="entries"
-								>
-									{tabView}
-									{totalItems > 0 &&
-										<>
-											<Divider />
-											{handlePagination(handleEntriesTabView)}
-										</>
-									}
-								</Tabs.TabPane>
-								<Tabs.TabPane
-									tab={
-										<Typography.Text strong>
-											<UpOutlined style={{ marginRight: 0 }} /> Up Voted Entries
-										</Typography.Text>
-									}
-									key="up-votes"
-								>
-									{tabView}
-									{totalItems > 0 &&
-										<>
-											<Divider />
-											{handlePagination(handleVotesTabView, 'up')}
-										</>
-									}
-								</Tabs.TabPane>
-								<Tabs.TabPane
-									tab={
-										<Typography.Text strong>
-											<DownOutlined style={{ marginRight: 0 }} /> Down Voted Entries
-										</Typography.Text>
-									}
-									key="down-votes"
-								>
-									{tabView}
-									{totalItems > 0 &&
-										<>
-											<Divider />
-											{handlePagination(handleVotesTabView, 'down')}
-										</>
-									}
-								</Tabs.TabPane>
-							</Tabs>
-						</Card>
+						<UserTabs username={match.params.username} />
 					</Col>
 				</Row>
 				<br/>
